vfs/cbe: initial VFS plugin using the CBE library
This plugin provides access to the virtual block-device of the CBE via
pseudo files.

There exist serveral directories, all serving a different purpose. The
'control' directory contains all files that manage the CBE, like a file
for setting the key as well as creating and discarding snapshots. The
'current' directory gives r/w access to the current working state of the
CBE. The 'snapshots' directory on the other hand only gives r/o access
to all other stats of the CBE.

The basic idea behind using different directories is to restrict the
access to different parts of the CBE. A trusted management component
may access the 'control' and 'snapshot' directories while the client,
e.g. a VM, has only access to the 'current' directory.