DIRTY run: add persistent storage run scripts (cbe fails)
The 'cbe_init' run script will use the 'cbe_block' component to
initialize a image file with a clean CBE state. The default image size
is 512MiB and can be changed by setting the environment variable
'CBE_IMAGE_SIZE' (the value is given in MiB). Half of the image is used
for the initial VBD and the 8th part for the FT.

The 'cbe_test_compare' run script will override the first 4MiB of the
VBD twice while checking (i.e., the data that was written is read back
again) each block one by one.